Move `const_evaluate_operand` to utils
