+++ xdsl/transforms/canonicalization_patterns/cf.py
@@ -9,7 +9,7 @@ from xdsl.pattern_rewriter import (
     op_type_rewrite_pattern,
 )
 from xdsl.rewriter import InsertPoint
-from xdsl.transforms.canonicalization_patterns.scf import const_evaluate_operand
+from xdsl.transforms.canonicalization_patterns.utils import const_evaluate_operand
 
 
 class AssertTrue(RewritePattern):
